+++ src/components/calcite-icon/calcite-icon.stories.ts
@@ -1,4 +1,4 @@
-import { withKnobs, select } from "@storybook/addon-knobs";
+import { select } from "@storybook/addon-knobs";
 import { darkBackground, iconNames, parseReadme, boolean } from "../../../.storybook/helpers";
 import readme from "./readme.md";
 
@@ -6,14 +6,13 @@ const notes = parseReadme(readme);
 
 export default {
   title: "components|Icon",
-  decorators: [withKnobs],
   parameters: { notes }
 };
 
 export const simple = () => `
   <calcite-icon
     icon="${select("icon", iconNames, iconNames[0])}"
-    scale="${select("size", ["s", "m", "l"], "m")}"
+    scale="${select("scale", ["s", "m", "l"], "m")}"
     ${boolean("filled", false)}
   ></calcite-icon>
 `;