+++ src/components/calcite-radio-button-group/calcite-radio-button-group.stories.ts
@@ -1,12 +1,11 @@
 import { storiesOf } from "@storybook/html";
-import { withKnobs, select } from "@storybook/addon-knobs";
+import { select } from "@storybook/addon-knobs";
 import { darkBackground, parseReadme, boolean } from "../../../.storybook/helpers";
 import readme from "./readme.md";
 
 const notes = parseReadme(readme);
 
 storiesOf("components|Radio Button Group", module)
-  .addDecorator(withKnobs)
   .add(
     "Light Theme",
     () => `