+++ src/components/calcite-tile-select-group/calcite-tile-select-group.tsx
@@ -3,7 +3,7 @@ import { Component, Host, h } from "@stencil/core";
 @Component({
   tag: "calcite-tile-select-group",
   styleUrl: "calcite-tile-select-group.scss",
-  shadow: true,
+  shadow: true
 })
 export class CalciteTileSelectGroup {
   render() {