+++ CHANGELOG.md
@@ -7,9 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Fixed
+
+- `calcite-pagination` - fixed bug with single page result sets
+
 ### Added
 
 - `calcite-tab-title` - now supports icons: `icon-start` and `icon-end` props have been added for explicit positioning of up to two icons.
+- `calcite-tabs` - tabs can now be positioned `above` (default) or `below` the tab content with the `position` prop
 - New component `calcite-rating`
 
 ### Updated